chore: move paypal status logic to a hook
