+++ packages/lib/src/components/PayPal/components/PaypalCreditComponent.tsx
@@ -1,8 +1,6 @@
 import { h } from 'preact';
-import { useCallback, useEffect, useRef, useState } from 'preact/hooks';
 
-import { ComponentMethodsRef } from '../../types';
-import { PayPalV6OnApproveData } from '../paypal-js-types';
+import { usePayPalStatus } from '../hooks/usePayPalStatus';
 import { PayPalCreditButton } from './PayPalCreditButton';
 import { PayPalProcessingSpinner } from './PayPalProcessingSpinner';
 import { PayPalSpinner } from './PayPalSpinner';
@@ -25,35 +23,11 @@ export const PaypalCreditComponent = ({
         countryCode?: string;
     }
 >) => {
-    const [status, setStatus] = useState('pending');
-
-    const paypalCreditComponentRef = useRef<ComponentMethodsRef>({
-        setStatus: setStatus
+    const { status, handleOnApprove } = usePayPalStatus({
+        paypalService,
+        onApprove,
+        setComponentRef
     });
-
-    useEffect(() => {
-        setComponentRef(paypalCreditComponentRef.current);
-    }, [setComponentRef]);
-
-    useEffect(() => {
-        paypalService
-            .isSdkLoaded()
-            .then(() => {
-                setStatus('ready');
-            })
-            .catch(() => {
-                // SDK failed to load, but we don't need to handle it here
-            });
-    }, [paypalService]);
-
-    const handleOnApprove = useCallback(
-        (data: PayPalV6OnApproveData) => {
-            setStatus('processing');
-            void onApprove(data);
-            return Promise.resolve();
-        },
-        [onApprove]
-    );
 
     if (status === 'pending') {
         return (
